Infer load balancer type on create
Rather than asking the user to explicitly specify if they want to
create an application or network load balancer, infer which one they
intend by the combination of ports supplied. By first validating that
the user does not supply a comingled set of TCP/HTTP ports, we can
infer the right type by the protocol of the any supplied port.