Eliminate desired status output
Not pulling its weight. This doesn't seem to communicate much as it is
mostly "Running" if it appears in the task list save for the 30 or so
seconds a task takes to shut down gracefully.

It would be more useful to derive states from the combination of
last and desired statuses like "Stopping" or "Draining." Unfortunately
ECS doesn't expose enough information to derive "Draining."