Create default security group for Fargate CLI
The default VPC security groups allows no ingress traffic by default. For
simplicity, for now we're going to create a single security group that
allows all inbound traffic. As we're running managed resources (i.e.
ELBs and containers on AWS Fargate), there is not much upside to forcing
the user to specify ports as part of running tasks and/or services at
the moment.